requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
boto3==1.20.5
botocore==1.23.5
cachetools==4.2.4
celery==5.2.0
certifi==2021.10.8
cffi==1.15.1
channels==3.0.4
channels-redis==3.4.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
contextlib2==21.6.0
cryptography==38.0.3
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-channels==0.7.0
django-cors-headers==3.10.0
django-extra-fields==3.0.2
django-filter==21.1
django-lint==2.0.4
django-ninja==0.19.1
django-ninja-extra==0.16.0
django-ninja-jwt==5.2.2
django-redis==5.0.0
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
fitparse==1.2.0
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gpxpy==1.5.0
h11==0.14.0
haversine==2.5.1
hiredis==2.0.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
injector==0.19.0
jmespath==0.10.0
kombu==5.2.1
logilab-astng==0.24.3
logilab-common==1.9.7
msgpack==1.0.4
mypy-extensions==0.4.3
ninja-schema==0.13.0
oauthlib==3.1.1
openpyxl==3.0.9
Pillow==8.4.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pydantic==1.10.2
PyJWT
pylint==0.28.0
pyOpenSSL==22.1.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.0.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.3.0
rsa==4.7.2
s3transfer==0.5.0
sentry-sdk==1.4.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.7
utm==0.7.0
uvicorn==0.19.0
vine==5.0.0
wcwidth==0.2.5
websockets==10.4
wrapt==1.13.3
zope.interface==5.5.1
django-extensions==3.2.1
pylint-quotes
black
eventsourcing-django==0.3.1
kafka-python==2.0.2
mysqlclient